Add ECC selection for 19.2 and above
Default is to prioritize secp256r1 as a curve, then other 256 bits, then
384 bits, and then the regular order (571 down to 160 bits)

This is all a bit complex, but is an attempt at balancing handshake
performance with speed; secp256r1 is the curve AWS ELBs always choose by
default and which we used to use and remains default; the rest allows to
pick curve according to our order of preference, based on whatever the
client contacting us might support.

The fix is backwards compatible to 19.1.1 and 19.1 at the very least, in
which case we fallback to existing requirements.