Allow overriding of default TLS versions.
Prior to this commit, the TLS versions 1.0, 1.1, and 1.2 were mandated
by default, and could not be overriden.

Times change and TLSv1.0 is often a thing you'd want to disable, and
even TLSv1.1. OTP-22 introduces initial (non-production) implementations
of TLSv1.3 as well, which users may want to enable.

This commit fixes this by moving the default versions to just being a
default (so that if they are not specified, you don't accidentally
enable even older TLS versions), but allowing them to be overriden in
snit:start_opts/6. This also allows per-server options to be set there,
creating safer or riskier endpoints on purpose (the same change can't be
done with SNI only).